Move order and checkout related discount methods to corresponding files
