+++ saleor/graphql/checkout/tests/benchmark/test_checkout_mutations.py
@@ -1297,7 +1297,7 @@ def test_add_checkout_lines_gift_discount_applies(
     }
 
     # when
-    with django_assert_num_queries(113):
+    with django_assert_num_queries(112):
         response = user_api_client.post_graphql(MUTATION_CHECKOUT_LINES_ADD, variables)
 
     # then
