+++ saleor/order/tests/test_fetch.py
@@ -34,7 +34,7 @@ def test_fetch_draft_order_lines_info(
     )
 
     # when
-    with django_assert_num_queries(11):
+    with django_assert_num_queries(12):
         lines_info = fetch_draft_order_lines_info(order)
 
     # then
