+++ saleor/tax/tests/test_checkout_calculations.py
@@ -7,7 +7,7 @@ from ...checkout.fetch import fetch_checkout_info, fetch_checkout_lines
 from ...checkout.utils import add_variant_to_checkout
 from ...core.prices import quantize_price
 from ...core.taxes import zero_taxed_money
-from ...discount.utils.promotion import (
+from ...discount.utils.checkout import (
     create_checkout_discount_objects_for_order_promotions,
 )
 from ...plugins.manager import get_plugins_manager
